+++ scripts/macros/2014/classFeatures/ranger/gloomStalker/dreadAmbusher.js
@@ -1,6 +1,8 @@
+import {workflowUtils} from '../../../../../utils.js';
+
 async function turnStart({trigger: {entity: item}}) {
     if (game.combat.round !== 1) return;
-    await item.use();
+    await workflowUtils.completeItemUse(item);
 }
 export let dreadAmbusher = {
     name: 'Dread Ambusher',