+++ scripts/macros/2014/classFeatures/warlock/fiend/darkOnesBlessing.js
@@ -1,9 +1,9 @@
-import {itemUtils} from '../../../../../utils.js';
+import {itemUtils, workflowUtils} from '../../../../../utils.js';
 
 async function late({trigger: {entity: item}, workflow}) {
     if (!workflow.hitTargets.size || !workflow.damageList) return;
     if (!workflow.damageList.some(i => i.oldHP > 0 && i.newHP === 0)) return;
-    await item.use();
+    await workflowUtils.completeItemUse(item);
 }
 export let darkOnesBlessing = {
     name: 'Dark One\'s Blessing',