+++ scripts/macros/2014/classFeatures/warlock/invocations/investmentOfTheChainMaster.js
@@ -1,4 +1,4 @@
-import {actorUtils, dialogUtils, effectUtils, genericUtils, itemUtils, socketUtils} from '../../../../../utils.js';
+import {actorUtils, dialogUtils, effectUtils, genericUtils, itemUtils, socketUtils, workflowUtils} from '../../../../../utils.js';
 async function hit({trigger: {entity: item, token}, workflow}) {
     if (!workflow.hitTargets.has(token)) return;
     let effect = effectUtils.getEffectByIdentifier(token.actor, 'summonedEffect');
@@ -9,7 +9,7 @@ async function hit({trigger: {entity: item, token}, workflow}) {
     let selection = await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: item.name}), {userId: socketUtils.firstOwner(sourceActor, true)});
     if (!selection) return;
     await actorUtils.setReactionUsed(sourceActor);
-    await item.use();
+    await workflowUtils.completeItemUse(item);
 }
 export let investmentOfTheChainMaster = {
     name: 'Eldritch Invocations: Investment of the Chain Master',